src/requirements.txt
zoomus==1.1.6
adafruit-circuitpython-neopixel
